更新至09.26（高度控制器更换/自动功能测试基本ok），基于no5 AC。

+++ ac_code/sw/airborne/firmwares/rotorcraft/nav_flight.h
@@ -28,7 +28,7 @@ enum Flight_State
 
 struct config_info
 {                          
-	uint8_t spray_concentration;  //unit=ml/m2
+	uint16_t spray_concentration;  //unit=ml/m2
 	uint8_t atomization_grade;  //grade 1/2/3...
 	float spray_height;     //unit=m
 	float spray_wide;       //unit=m
